2019/pyproject.toml
[tool.poetry]
name = "aoc-2019"
version = "0.1.0"
description = ""
authors = ["Treelar <t.rothled@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
networkx = "^2.8.4"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
